Fixed delete trigger, added state flow to record edit

+++ app/src/main/java/com/example/paymeback/data/OfflineRecordsRepository.kt
@@ -8,7 +8,7 @@ class OfflineRecordsRepository @Inject constructor(
 ) : RecordsRepository {
     override fun getAllRecordsStream(): Flow<List<Record>> = recordDao.getAllRecords()
 
-    override fun getRecordStream(id: Int): Flow<Record?> = recordDao.getRecord(id)
+    override fun getRecordStream(id: Long): Flow<Record?> = recordDao.getRecord(id)
 
     override fun getRecordWithPaymentsStream(id: Long): Flow<RecordWithPayments?> =
         recordDao.getRecordWithPayments(id)
